+++ R/reexport-svBase.R
@@ -1,0 +1,3 @@
+#' @importFrom svBase f_
+#' @export
+svBase::f_